+++ Software/MATLAB/app/system/+matlab/+compiler/+mlspark/@Dataset/na.m
@@ -1,0 +1,9 @@
+function dfnf = na(obj)
+    % na DataFrameNaFunctions object
+
+    % Copyright 2022 MathWorks, Inc.
+
+
+    dfnf = matlab.compiler.mlspark.DataFrameNaFunctions(obj.dataset.na());
+
+end %function