+++ Software/MATLAB/test/fixtures/f_inT2_outT1.m
@@ -1,0 +1,9 @@
+function OT = f_inT2_outT1(IT1, IT2)
+    % f_inT2_outT1 Argument checking function
+    
+    % Copyright 2022 MathWorks, Inc.
+   
+    OT = normalize(IT1);
+    head(IT2)
+
+end